+++ __docs__/components/gallery/tiles/one-pane-dialog-tile.tsx
@@ -8,6 +8,7 @@ import {Body} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
 import {CommonTileProps} from "../types";
+import {modalPositionerStyle} from "../../../wonder-blocks-modal/modal-story-utils";
 
 export default function OnePaneDialogTile(props: CommonTileProps) {
     return (
@@ -51,25 +52,9 @@ export default function OnePaneDialogTile(props: CommonTileProps) {
 const mobile = "@media (max-width: 1023px)";
 
 const localStyles = StyleSheet.create({
-    modalPositioner: {
-        // Checkerboard background
-        backgroundImage:
-            "linear-gradient(45deg, #ccc 25%, transparent 25%), linear-gradient(-45deg, #ccc 25%, transparent 25%), linear-gradient(45deg, transparent 75%, #ccc 75%), linear-gradient(-45deg, transparent 75%, #ccc 75%)",
-        backgroundSize: "20px 20px",
-        backgroundPosition: "0 0, 0 10px, 10px -10px, -10px 0px",
-
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "center",
-
-        position: "absolute",
-        left: 0,
-        right: 0,
-        top: 0,
-        bottom: 0,
-    },
+    modalPositioner: modalPositionerStyle,
     previewSizer: {
-        minHeight: 500,
+        minBlockSize: 500,
         width: "100%",
 
         [mobile]: {