+++ __docs__/wonder-blocks-modal/one-pane-dialog.stories.tsx
@@ -20,6 +20,7 @@ import ComponentInfo from "../components/component-info";
 import OnePaneDialogArgTypes from "./one-pane-dialog.argtypes";
 import {allModes} from "../../.storybook/modes";
 import {reallyLongText} from "../components/text-for-testing";
+import {modalPositionerStyle} from "./modal-story-utils";
 
 export default {
     title: "Packages / Modal / OnePaneDialog",
@@ -570,25 +571,9 @@ const styles = StyleSheet.create({
         alignItems: "center",
         justifyContent: "center",
     },
-    modalPositioner: {
-        // Checkerboard background
-        backgroundImage:
-            "linear-gradient(45deg, #ccc 25%, transparent 25%), linear-gradient(-45deg, #ccc 25%, transparent 25%), linear-gradient(45deg, transparent 75%, #ccc 75%), linear-gradient(-45deg, transparent 75%, #ccc 75%)",
-        backgroundSize: "20px 20px",
-        backgroundPosition: "0 0, 0 10px, 10px -10px, -10px 0px",
-
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "center",
-
-        position: "absolute",
-        left: 0,
-        right: 0,
-        top: 0,
-        bottom: 0,
-    },
+    modalPositioner: modalPositionerStyle,
     previewSizer: {
-        minHeight: 600,
+        minBlockSize: 600,
         width: "100%",
     },
     row: {