+++ __docs__/wonder-blocks-tokens/semantic-color-groups.stories.tsx
@@ -254,6 +254,6 @@ const styles = StyleSheet.create({
     },
 
     banner: {
-        marginBottom: sizing.size_320,
+        marginBlockEnd: sizing.size_320,
     },
 });