+++ packages/wonder-blocks-tabs/src/components/tab.tsx
@@ -114,7 +114,7 @@ export const styles = StyleSheet.create({
         display: "flex",
         alignItems: "center",
         textWrap: "nowrap",
-        backgroundColor: "transparent",
+        backgroundColor: semanticColor.core.transparent,
         border: "none",
         margin: 0,
         padding: 0,
@@ -130,9 +130,9 @@ export const styles = StyleSheet.create({
         ":after": {
             content: "''",
             position: "absolute",
-            left: 0,
-            right: 0,
-            bottom: `calc(${bottomSpacing} * -1)`,
+            insetInlineStart: 0,
+            insetInlineEnd: 0,
+            insetBlockEnd: `calc(${bottomSpacing} * -1)`,
         },
         // Only apply hover styles to tabs that are not selected
         [":hover:not([aria-selected='true'])" as any]: {