+++ packages/wonder-blocks-switch/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-switch
 
+## 3.4.1
+
+### Patch Changes
+
+- 40cb70f: Add `require-logical-properties-for-rtl` ESLint rule to `eslint-plugin-wonder-blocks` recommended config, and migrate all Wonder Blocks component source files to use CSS logical properties for improved RTL layout support.
+- 40cb70f: Enable RTL logical-properties ESLint rule in recommended config
+- Updated dependencies [40cb70f]
+- Updated dependencies [40cb70f]
+    - @khanacademy/wonder-blocks-core@12.4.4
+    - @khanacademy/wonder-blocks-icon@5.3.15
+
 ## 3.4.0
 
 ### Minor Changes